Use temp and rewrite function with min and max
Rewrite function to make it more simple. Add the checks if number is
integer and check if the numbers are below 0. Then add temp variable
to make the function able to switch the value from min with the value of
max if min is bigger in the beginning. Last write the more simple for
loop again, same as before but more readable and simple